Made system exit explicitly if input data is invalid
Decided that if the input data is invalid, i.e. the grid size is less than zero or the live cells listed are not within the supplied grid size, then the program should fail with an appropriate message.
