+++ conways2/src/com/ness/Validator.java
@@ -11,23 +11,5 @@ class Validator {
         this.maxGridSize = maxGridSize;
     }
 
-    boolean validInputString(String s) {
-        Pattern numberPair = Pattern.compile("\\d*\\s\\d*");
-        Matcher matcher = numberPair.matcher(s);
-        return matcher.matches();
-    }
 
-    boolean validGridSize(int[] gridSize) {
-        if (gridSize[0] <=0 || gridSize[1] <= 0) {
-            return false;
-        }
-        return (gridSize[0] < maxGridSize && gridSize[1] < maxGridSize);
-    }
-
-    boolean validCellLocation(int[] cellLocation, boolean[][] grid) {
-        if (cellLocation[0] <= 0 || cellLocation[1] <= 0) {
-            return false;
-        }
-        return (cellLocation[1] <= grid.length && cellLocation[0] <= grid[0].length);
-    }
 }
