+++ conways2/src/messages.properties
@@ -1,1 +1,2 @@
 welcome=Welcome to Conway's Game of Life!
+invalid=Invalid input found. System is exiting.
